feat: add factory methods for webdriver waits
Especially with Selenium refining its API, it makes sense to hide
these details behind Galenium.

Also some javadoc.
