+++ modules/icing/src/main/java/io/wcm/qa/glnm/interaction/Wait.java
@@ -19,7 +19,8 @@
  */
 package io.wcm.qa.glnm.interaction;
 
-import java.time.Duration;
+import static io.wcm.qa.glnm.webdriver.WebDriverManagement.getWait;
+
 import java.util.function.Function;
 
 import org.openqa.selenium.JavascriptExecutor;
@@ -33,7 +34,6 @@ import org.slf4j.LoggerFactory;
 import io.wcm.qa.glnm.exceptions.GaleniumException;
 import io.wcm.qa.glnm.sampling.CanCache;
 import io.wcm.qa.glnm.selectors.base.Selector;
-import io.wcm.qa.glnm.util.GaleniumContext;
 import io.wcm.qa.glnm.verification.base.Verifiable;
 import io.wcm.qa.glnm.verification.base.Verification;
 import io.wcm.qa.glnm.verification.element.InvisibilityVerification;
@@ -47,7 +47,6 @@ import io.wcm.qa.glnm.verification.element.VisibilityVerification;
  */
 public final class Wait {
 
-  private static final int DEFAULT_NUMBER_OF_POLLS_PER_CALL = 5;
   private static final int DEFAULT_TIMEOUT = 1;
 
   private static final Logger LOG = LoggerFactory.getLogger(Wait.class);
@@ -57,50 +56,45 @@ public final class Wait {
   }
 
   /**
-   * Waits for  {@link io.wcm.qa.glnm.verification.base.Verifiable} or  {@link io.wcm.qa.glnm.verification.base.Verification}.
+   * Waits for {@link io.wcm.qa.glnm.verification.base.Verifiable} or
+   * {@link io.wcm.qa.glnm.verification.base.Verification}.
    *
    * @param condition to wait for
+   * @since 1.0.0
    */
   public static void forCondition(Verifiable condition) {
-    int timeOut = DEFAULT_TIMEOUT;
-    forCondition(condition, timeOut);
+    forCondition(condition, DEFAULT_TIMEOUT);
   }
 
   /**
-   * Waits for  {@link io.wcm.qa.glnm.verification.base.Verifiable} or  {@link io.wcm.qa.glnm.verification.base.Verification}.
+   * Waits for {@link io.wcm.qa.glnm.verification.base.Verifiable} or
+   * {@link io.wcm.qa.glnm.verification.base.Verification}.
    *
    * @param condition to wait for
    * @param timeout how many seconds to wait
+   * @since 1.0.0
    */
   public static void forCondition(Verifiable condition, int timeout) {
-    int pollingInterval = getPollingIntervalForTimeout(timeout);
-    forCondition(condition, timeout, pollingInterval);
+    forCondition(condition, getWait(timeout));
   }
 
   /**
-   * Waits for  {@link io.wcm.qa.glnm.verification.base.Verifiable} or  {@link io.wcm.qa.glnm.verification.base.Verification}.
+   * Waits for {@link io.wcm.qa.glnm.verification.base.Verifiable} or
+   * {@link io.wcm.qa.glnm.verification.base.Verification}.
    *
    * @param condition to wait for
    * @param timeOut how many seconds to wait
    * @param pollingInterval how many milliseconds between attempts
+   * @since 1.0.0
    */
   public static void forCondition(Verifiable condition, int timeOut, int pollingInterval) {
-    WebDriverWait wait = getWait(timeOut, pollingInterval);
-    VerifiableExpectedCondition verifiableCondition = new VerifiableExpectedCondition(condition);
-    try {
-      wait.until(verifiableCondition);
-    }
-    catch (TimeoutException ex) {
-      if (verifiableCondition.isVerification()) {
-        Verification verification = (Verification)verifiableCondition.getVerifiable();
-        throw new GaleniumException(verification.getMessage(), ex);
-      }
-      throw ex;
-    }
+    forCondition(condition, getWait(timeOut, pollingInterval));
   }
 
   /**
    * Wait for domReady for a maximum of one second.
+   *
+   * @since 1.0.0
    */
   public static void forDomReady() {
     forDomReady(DEFAULT_TIMEOUT);
@@ -110,6 +104,7 @@ public final class Wait {
    * Wait for "DOM ready" for a maximum of seconds specified by parameter.
    *
    * @param timeOutInSeconds how long to wait for "DOM ready"
+   * @since 1.0.0
    */
   public static void forDomReady(int timeOutInSeconds) {
     WebDriverWait wait = getWait(timeOutInSeconds);
@@ -120,6 +115,7 @@ public final class Wait {
    * Wait for element to be invisible or not in page.
    *
    * @param selector identifies element
+   * @since 1.0.0
    */
   public static void forInvisibility(Selector selector) {
     forCondition(new InvisibilityVerification(selector), DEFAULT_TIMEOUT);
@@ -130,6 +126,7 @@ public final class Wait {
    *
    * @param selector identifies element
    * @param timeout how many seconds to wait
+   * @since 1.0.0
    */
   public static void forInvisibility(Selector selector, int timeout) {
     forCondition(new InvisibilityVerification(selector), timeout);
@@ -139,6 +136,7 @@ public final class Wait {
    * Wait for element to display text from expected properties.
    *
    * @param selector identifies element
+   * @since 1.0.0
    */
   public static void forText(Selector selector) {
     forText(selector, null, DEFAULT_TIMEOUT);
@@ -149,6 +147,7 @@ public final class Wait {
    *
    * @param selector identifies element
    * @param timeout how many seconds to wait
+   * @since 1.0.0
    */
   public static void forText(Selector selector, int timeout) {
     forText(selector, null, timeout);
@@ -159,6 +158,7 @@ public final class Wait {
    *
    * @param selector identifies element
    * @param text to match
+   * @since 1.0.0
    */
   public static void forText(Selector selector, String text) {
     forText(selector, text, DEFAULT_TIMEOUT);
@@ -170,6 +170,7 @@ public final class Wait {
    * @param selector identifies element
    * @param text to match
    * @param timeout how many seconds to wait
+   * @since 1.0.0
    */
   public static void forText(Selector selector, String text, int timeout) {
     forCondition(new TextVerification(selector, text), timeout);
@@ -179,6 +180,7 @@ public final class Wait {
    * Load URL and wait for it to be loaded.
    *
    * @param url to load
+   * @since 1.0.0
    */
   public static void forUrl(String url) {
     forUrl(url, DEFAULT_TIMEOUT);
@@ -189,6 +191,7 @@ public final class Wait {
    *
    * @param url to load
    * @param timeOutInSeconds how long to wait for URL to be current
+   * @since 1.0.0
    */
   public static void forUrl(String url, int timeOutInSeconds) {
     LOG.trace("waiting for URL: '" + url + "'");
@@ -201,6 +204,7 @@ public final class Wait {
    * Wait for element to be visible.
    *
    * @param selector identifies element
+   * @since 1.0.0
    */
   public static void forVisibility(Selector selector) {
     forCondition(new VisibilityVerification(selector), DEFAULT_TIMEOUT);
@@ -211,25 +215,24 @@ public final class Wait {
    *
    * @param selector identifies element
    * @param timeout how many seconds to wait
+   * @since 1.0.0
    */
   public static void forVisibility(Selector selector, int timeout) {
     forCondition(new VisibilityVerification(selector), timeout);
   }
 
-  private static int getPollingIntervalForTimeout(int timeoutInSeconds) {
-    int timeoutInMillis = timeoutInSeconds * 1000;
-    return timeoutInMillis / DEFAULT_NUMBER_OF_POLLS_PER_CALL;
-  }
-
-  private static WebDriverWait getWait(int timeOutInSeconds) {
-    return getWait(timeOutInSeconds, getPollingIntervalForTimeout(timeOutInSeconds));
-  }
-
-  private static WebDriverWait getWait(int timeOutInSeconds, int pollingInterval) {
-    Duration polling = Duration.ofMillis(pollingInterval);
-    Duration timeout = Duration.ofSeconds(timeOutInSeconds);
-    WebDriver driver = GaleniumContext.getDriver();
-    return new WebDriverWait(driver, timeout, polling);
+  private static void forCondition(Verifiable condition, WebDriverWait wait) {
+    VerifiableExpectedCondition verifiableCondition = new VerifiableExpectedCondition(condition);
+    try {
+      wait.until(verifiableCondition);
+    }
+    catch (TimeoutException ex) {
+      if (verifiableCondition.isVerification()) {
+        Verification verification = (Verification)verifiableCondition.getVerifiable();
+        throw new GaleniumException(verification.getMessage(), ex);
+      }
+      throw ex;
+    }
   }
 
   private static final class VerifiableExpectedCondition implements Function<WebDriver, Boolean> {
